Refactor to microservices architecture with independent services
