+++ microservices/auth-service/src/main/java/com/perro/auth/model/Usuario.java
@@ -4,9 +4,6 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Column;
 import org.springframework.data.relational.core.mapping.Table;
 
-import lombok.Data;
-
-@Data
 @Table("usuarios")
 public class Usuario {
 
@@ -18,4 +15,23 @@ public class Usuario {
 
     @Column("password")
     private String password;
+
+    public Long getId() {
+        return id;
+    }
+    public void setId(Long id) {
+        this.id = id;
+    }
+    public String getEmail() {
+        return email;
+    }
+    public void setEmail(String email) {
+        this.email = email;
+    }
+    public String getPassword() {
+        return password;
+    }
+    public void setPassword(String password) {
+        this.password = password;
+    }
 }
