+++ microservices/pet-service/src/main/java/com/perro/pet/service/PerroService.java
@@ -12,37 +12,45 @@ import reactor.core.publisher.Mono;
 @Service
 @RequiredArgsConstructor
 public class PerroService {
-    
-    private final PerroRepository repository;
-    
-    public Flux<Perro> listarPorUsuario(Long usuarioId) {
-        return repository.findByUsuarioId(usuarioId);
+
+    private final PerroRepository repo;
+
+    public Flux<Perro> listar() {
+        return repo.findAll();
     }
-    
-    public Mono<Perro> registrar(Perro perro) {
-        return repository.save(perro);
+
+    public Mono<Perro> registrar(Perro h) {
+        return repo.save(h);
     }
-    
-    public Mono<Perro> actualizar(Long id, Long usuarioId, Perro perro) {
-        return repository.findByIdAndUsuarioId(id, usuarioId)
-                .flatMap(existingPerro -> {
-                    perro.setId(id);
-                    perro.setUsuarioId(usuarioId);
-                    return repository.save(perro);
-                });
-    }
-    
-    public Mono<Void> eliminarPorUsuario(Long id, Long usuarioId) {
-        return repository.findByIdAndUsuarioId(id, usuarioId)
-                .flatMap(perro -> repository.deleteById(id))
-                .then();
-    }
-    
-    public Mono<Perro> obtenerPorId(Long id, Long usuarioId) {
-        return repository.findByIdAndUsuarioId(id, usuarioId);
-    }
-    
+
     public Mono<Void> eliminar(Long id) {
-        return repository.deleteById(id);
+        return repo.deleteById(id);
+    }
+
+    public Mono<Perro> obtenerPorId(Long id) {
+        return repo.findById(id);
+    }
+
+    public Flux<Perro> listarPorUsuario(Long usuarioId) {
+        return repo.findAllByUsuarioId(usuarioId);
+    }
+
+    public Mono<Perro> actualizar(Long id, Long usuarioId, Perro nuevoPerro) {
+        return repo.findByIdAndUsuarioId(id, usuarioId)
+            .flatMap(perro -> {
+                perro.setNombre(nuevoPerro.getNombre());
+                perro.setRaza(nuevoPerro.getRaza());
+                perro.setTamano(nuevoPerro.getTamano());
+                perro.setUbicacion(nuevoPerro.getUbicacion());
+                perro.setComportamiento(nuevoPerro.getComportamiento());
+                perro.setDuenio(nuevoPerro.getDuenio());
+                perro.setEdad(nuevoPerro.getEdad());
+                return repo.save(perro);
+            });
+    }
+
+    public Mono<Void> eliminarPorUsuario(Long id, Long usuarioId) {
+        return repo.findByIdAndUsuarioId(id, usuarioId)
+            .flatMap(perro -> repo.deleteById(perro.getId()));
     }
 }
