+++ microservices/pet-service/src/main/resources/application.yml
@@ -1,22 +1,23 @@
 server:
-  port: 8082
+  port: 8081
 
 spring:
   application:
     name: pet-service
+  datasource:
+    url: jdbc:postgresql://localhost:5432/mascotas_db
+    username: postgres
+    password: postgres
+    driver-class-name: org.postgresql.Driver
   r2dbc:
-    url: r2dbc:mysql://localhost:3306/pet_db
-    username: root
-    password: admin
-  sql:
-    init:
-      mode: always
-
-eureka:
-  client:
-    service-url:
-      defaultZone: http://localhost:8761/eureka/
+    url: r2dbc:postgresql://localhost:5432/mascotas_db
+    username: postgres
+    password: postgres
+  cloud:
+    discovery:
+      enabled: true
+      service-id: pet-service
 
 logging:
   level:
-    org.springframework.r2dbc: DEBUG
+    com.perro.pet: DEBUG
