--- src/main/java/com/perro/mascotas/MascotasApplication.java
@@ -1,13 +1,0 @@
-package com.perro.mascotas;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MascotasApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(MascotasApplication.class, args);
-	}
-
-}
