--- src/main/java/com/perro/mascotas/model/Usuario.java
@@ -1,37 +1,0 @@
-package com.perro.mascotas.model;
-
-import org.springframework.data.annotation.Id;
-import org.springframework.data.relational.core.mapping.Column;
-import org.springframework.data.relational.core.mapping.Table;
-
-@Table("usuarios")
-public class Usuario {
-
-    @Id
-    private Long id;
-
-    @Column("email")
-    private String email;
-
-    @Column("password")
-    private String password;
-
-    public Long getId() {
-        return id;
-    }
-    public void setId(Long id) {
-        this.id = id;
-    }
-    public String getEmail() {
-        return email;
-    }
-    public void setEmail(String email) {
-        this.email = email;
-    }
-    public String getPassword() {
-        return password;
-    }
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
