--- src/main/resources/application.yml
@@ -1,14 +1,0 @@
-spring:
-  r2dbc:
-    url: r2dbc:mysql://localhost:3306/perrosbd
-    username: root
-    password: admin
-  sql:
-    init:
-      mode: always
-
-logging:
-  level:
-    org.springframework.r2dbc: DEBUG
-server:
-  port: 8080
