--- src/test/java/com/perro/mascotas/MascotasApplicationTests.java
@@ -1,13 +1,0 @@
-package com.perro.mascotas;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MascotasApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
